Split the resource file and translate in a thread Changed the status type to be a float instead of an int Changed from using googleTrans package to Translators pacakge

+++ translate/models.py
@@ -24,5 +24,5 @@ class Session(models.Model):
 
 class FileStatus(models.Model):
     session = models.ForeignKey(Session, on_delete=models.CASCADE)
-    status = models.IntegerField()
+    status = models.FloatField()
